Ag broadcast reduce3 smoke tests (#10159)
* add new smoke tests
fix scalar index when using the indexing api
fix usage of index2coords by migrating shape information
to shape::shapeOf
fix usage of coords2index by migrating accidental shapeOf invocations to strides

* Fix reduce3 buffer() calls in op execution

* fix test that didn't account for overlapping views

* remove print statements

+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/shape/bp/TileBp.java
@@ -30,9 +30,9 @@ import java.util.*;
 
 public class TileBp extends DynamicCustomOp {
 
-    private int[] repeat;
+    private long[] repeat;
 
-    public TileBp(SameDiff sameDiff, SDVariable in, SDVariable grad, int[] repeat) {
+    public TileBp(SameDiff sameDiff, SDVariable in, SDVariable grad, long[] repeat) {
         super(null,sameDiff, new SDVariable[]{in, grad}, false);
         this.repeat = repeat;
         addArguments();
